feat(dbt): project-wide Runs view + unified breadcrumb bar
Add a Transforms "Runs" view listing every run for a project — including
ad-hoc agent (dbt_run_model) and editor "Run model" runs that have no
jobId and previously surfaced only in chat. Extracts the shared
run-history + live-log panel from DbtJobView into DbtRunHistory, adds a
project-root "Runs" tree node + dbt-runs tab, self-polls while runs are
active, and lets the chat run card jump straight to it.

Also collapse the dbt editor's redundant "project / path" sub-header into
the standard breadcrumb: the breadcrumb now includes the project name,
and the dbt file editor + Runs view render their own EntityBreadcrumbs
with actions in the trailing slot (matching app-file / table-data).
